+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Changelog
 
 - d20 alternative can now be passed into Actor#rollX and Item#useAttack API (eg "2d20kh")
+- Improve actor sheet rendering performance
 
 ## 0.76.2